refactor Bundle into Graph and Bundle
