+++ src/ast/nodes/IfStatement.ts
@@ -52,7 +52,7 @@ export default class IfStatement extends Statement {
 
 	initialiseChildren (parentScope: Scope) {
 		super.initialiseChildren(parentScope);
-		if (this.module.bundle.treeshake) {
+		if (this.module.graph.treeshake) {
 			this.testValue = this.test.getValue();
 
 			if (this.testValue === UNKNOWN_VALUE) {
@@ -71,7 +71,7 @@ export default class IfStatement extends Statement {
 	}
 
 	render (code: MagicString, es: boolean) {
-		if (this.module.bundle.treeshake) {
+		if (this.module.graph.treeshake) {
 			if (this.testValue === UNKNOWN_VALUE) {
 				super.render(code, es);
 			} else {
