+++ src/ast/scopes/GlobalScope.ts
@@ -1,7 +1,7 @@
 import GlobalVariable from '../variables/GlobalVariable';
 import Scope from './Scope';
 
-export default class BundleScope extends Scope {
+export default class GlobalScope extends Scope {
 	parent: void;
 
 	findVariable (name: string) {
