+++ src/finalisers/shared/getExportBlock.ts
@@ -19,7 +19,7 @@ export default function getExportBlock (
 			if (name[0] === '*') {
 				// export all from external
 				const id = name.slice(1);
-				const module = <ExternalModule> bundle.moduleById.get(id);
+				const module = <ExternalModule> bundle.graph.moduleById.get(id);
 
 				return `Object.keys(${
 					module.name
