+++ src/finalisers/shared/getInteropBlock.ts
@@ -8,13 +8,13 @@ export default function getInteropBlock (bundle: Bundle, options: OutputOptions)
 				return null;
 
 			if (module.exportsNamespace) {
-				return `${bundle.varOrConst} ${module.name}__default = ${
+				return `${bundle.graph.varOrConst} ${module.name}__default = ${
 					module.name
 					}['default'];`;
 			}
 
 			if (module.exportsNames) {
-				return `${bundle.varOrConst} ${module.name}__default = 'default' in ${
+				return `${bundle.graph.varOrConst} ${module.name}__default = 'default' in ${
 					module.name
 					} ? ${module.name}['default'] : ${module.name};`;
 			}
